Camera Savving images :D
I finaly got the camera to work! Now i can take a picture and it saves
it to picture/GeoSpot the opens it for the user to crop the picture that
will be sent to the online server (ill work on that 2 but not right now)

I hope in the next update ill have a new UI for when a user is logged
in.